Correct main entry point
Since 2.8.0, I have this error:

```
This dependency was not found:

* @yaireo/tagify in ./assets/js/admin.js
```

I guess it's because the main file is not found?

(Please, be kind with me! :pray:  It's my first PR on a pure javascript library and I don't know yet if I should be proud or ashamed of it! :p )